requirement.txt
asgiref             3.8.1
certifi             2024.2.2
distlib             0.3.8
Django              5.0.4
djangorestframework 3.15.1
filelock            3.13.1
pip                 23.3.2
pipenv              2023.12.0
platformdirs        4.2.0
setuptools          69.0.3
sqlparse            0.5.0
tzdata              2024.1
virtualenv          20.25.0